+++ db/schema.rb
@@ -13,7 +13,7 @@
 ActiveRecord::Schema.define(version: 2019_05_27_021028) do
 
   create_table "groups", id: :integer, unsigned: true, options: "ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci", force: :cascade do |t|
-    t.integer "chat_id", null: false
+    t.bigint "chat_id", null: false
     t.string "name", null: false
   end
